==> 02-data.Rmd <==
# Data Sources

## NYPD Arrests Data 

*(Year to Date)*

*Source: https://data.cityofnewyork.us/Public-Safety/NYPD-Arrest-Data-Year-to-Date-/uip8-fykc*

**Description**: This is a breakdown of every arrest effected in NYC by the NYPD during the current year. This data is manually extracted every quarter and reviewed by the Office of Management Analysis and Planning. Each record represents an arrest effected in NYC by the NYPD and includes information about the type of crime, the location and time of enforcement. In addition, information related to suspect demographics is also included. 

**Data Provided By**: New York Police Department

**Total records**: 140,413

**Issues with the data**: 

 - This data only contains information for 2020. Having more historic data would help in carrying out a more successful time-wise analysis. 
 - It would also be useful to have information about victims associated with these arrests.
 - Another useful feature could be county names, adding another level of location information. Drilling down to county level might result in interesting insights as well.

**Type of variables**: This can be seen below. There are 19 columns and their data types are noted below.

```{r}
df_arrest_ytd <- readr::read_csv("datasets/NYPD_Arrest_Data__Year_to_Date_.csv")

str(df_arrest_ytd)
```

## NYPD Hate Crimes

*Source: https://data.cityofnewyork.us/Public-Safety/NYPD-Hate-Crimes/bqiq-cu78*
**Description**: Dataset containing confirmed hate crime incidents in New York City. A hate crime is a traditional offense like murder, arson, or vandalism with an added element of bias. For the purposes of collecting statistics, the FBI has defined a hate crime as a “criminal offense against a person or property motivated in whole or in part by an offender’s bias against a race, religion, disability, sexual orientation, ethnicity, gender, or gender identity.”  

**Data Provided By**: New York Police Department

**Total records**: 728

**Issues with the data**: County names are historical.
```{r}
df_hate_crimes <- readr::read_csv("datasets/NYPD_Hate_Crimes.csv")
```

**Type of variables**: 
```{r}
str(df_hate_crimes)
```

## NYC Shooting Data 

### NYPD Shooting Incident Data 

*Source: https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Year-To-Date-/5ucz-vwe8*

**Description**: 
 
 - This is a breakdown of every shooting incident that occurred in NYC during the current calendar year. (2020) This data is manually extracted every quarter and reviewed by the Office of Management Analysis and Planning before being posted on the NYPD website.
Each record represents a shooting incident in NYC and includes information about the event, the location and time of occurrence. 
 - In addition, information related to suspect and victim demographics is also included. For example, has information about the sex, Age, Race of the victim.
 
**Data Provided By**: New York Police Department

**Total records**: 1,942

**Issues with the data**: Contains only data for 2020. Searched for alternatives for more historical information.

**Type of variables**: 

```{r}
df_shooting_ytd <- readr::read_csv("datasets/NYPD_Shooting_Incident_Data__Year_To_Date_.csv")
covid_date<- as.Date("2020-04-01", format = "%Y-%m-%d")
str(df_shooting_ytd)
```

### NYPD Shooting Incident Historic Data

*Source: https://data.cityofnewyork.us/Public-Safety/NYPD-Shooting-Incident-Data-Historic-/833y-fsy8*

**Description**:

 - This is a breakdown of every shooting incident that occurred in NYC going back to 2006 through the end of the previous calendar year.(2019) This data is manually extracted every quarter and reviewed by the Office of Management Analysis and Planning before being posted on the NYPD website. Each record represents a shooting incident in NYC and includes information about the event, the location and time of occurrence. In addition, information related to suspect and victim demographics is also included.
 
**Data Provided By**: New York Police Department

**Total records**: 21,626

**Issues with the data**: Contains shootings starting from 2006 but the information of other years is too much to display together.

**Type of variables**: 

```{r}
df_shooting_old <- readr::read_csv("datasets/NYPD_Shooting_Incident_Data__Historic_.csv")
str(df_shooting_old)
```
 
## FBI Human Trafficking Data

*Source: https://crime-data-explorer.fr.cloud.gov/downloads-and-docs*

**Description**: The FBI began accepting data on human trafficking from states in January 2013. Human trafficking includes offenses related to commercial sex acts and involuntary servitude. Human Trafficking data available through the Crime Data Explorer include offenses and arrests recorded by state and local agencies that currently have the ability to report this crime to the national UCR Program. The UCR Program began collecting Human Trafficking data in 2013. The initial Human Trafficking publication was in 2013. 

**Data Provided By**: Federal Bureau Of Investigation

**Total records**: 3,335

**Issues with the data**: 

 - It would be useful to obtain more detailed information about perpetrators and involved criminal groups.
 - More specific time related information such as date and time may also be useful.

**Type of variables**: This can be seen below. There are 19 columns and their data types are noted below.

```{r}
df_fbi_ht <- readr::read_csv("datasets/HT_2013_2019.csv")
str(df_fbi_ht)
```

*Many other important datasets can be found on this website.*

## FBI Drug Arrests Data

This dataset contains monthly number of arrests for drug abuse violations reported by participating law enforcement agencies from 1995–2016. The arrests are by offense and broken down by age and sex or age and race.

*(Reported Number of Drug Arrests)*

*Source: https://crime-data-explorer.fr.cloud.gov/downloads-and-docs*

**Data Provided By**: FBI

**Total records**: 23

**Issues with the data**: The biggest issue with this dataset is it's limited size. There is no state-wise information, or month-wise information, which would helped deepen our analysis scope.

```{r}
df_fbi_drug_arrests <- readr::read_csv("datasets/arrests_national_drug.csv")
```
**Type of variables**: 
```{r}
str(df_fbi_drug_arrests)
```

## NYC Park Crime Data

*Source: https://www1.nyc.gov/site/nypd/stats/crime-statistics/park-crime-stats.page*

**Description**: This is a dataset containing crimes in the parks of New York City. This dataset is gleaned from quarterly reports from the NYPD. The data transformation is detailed in the cleaning file. 

**Data Provided By**: New York Police Department

**Total records**: 27696

**Issues with the data**: None

```{r}
df_parks<- readr::read_csv("datasets/parks/finalpark.csv")
```
**Type of variables**: 
```{r}
str(df_parks)
```
