==> 03-cleaning.Rmd <==
# Data Transformation

```{r}
library("naniar")
library("ggplot2")
library("dplyr")
library("tidyr")
library("lubridate")
library("ggpubr")
library("data.table")
```

## NYPD Arrests

In order to use arrest dates information, we are converting this feature to Date format.

```{r, echo=TRUE}
df_arrest_ytd$ARREST_DATE <-as.Date(df_arrest_ytd$ARREST_DATE, format = "%m/%d/%Y")
```

As per this mapping given by NYPD - B(Bronx), S(Staten Island), K(Brooklyn), M(Manhattan), Q(Queens) - we are updating the borough names for a clearer understanding.

```{r, echo=TRUE}
df_arrest_ytd$ARREST_BORO[df_arrest_ytd$ARREST_BORO == "B"] <- "Bronx"
df_arrest_ytd$ARREST_BORO[df_arrest_ytd$ARREST_BORO == "S"] <- "Staten Island"
df_arrest_ytd$ARREST_BORO[df_arrest_ytd$ARREST_BORO == "K"] <- "Brooklyn"
df_arrest_ytd$ARREST_BORO[df_arrest_ytd$ARREST_BORO == "M"] <- "Manhattan"
df_arrest_ytd$ARREST_BORO[df_arrest_ytd$ARREST_BORO == "Q"] <- "Queens"
```

This data also contains certain duplicate offenses which are categorized separately.

```{r}

# df_arrest_ytd %>% filter(OFNS_DESC %in% "KIDNAPPING AND RELATED OFFENSES")
# df_arrest_ytd %>% filter(OFNS_DESC %in% "KIDNAPPING")
# df_arrest_ytd %>% filter(OFNS_DESC %in% "ADMINISTRATIVE CODES")

```

```{r echo=TRUE}

df_arrest_ytd$OFNS_DESC[df_arrest_ytd$OFNS_DESC == "KIDNAPPING AND RELATED OFFENSES"] <- "KIDNAPPING & RELATED OFFENSES"
df_arrest_ytd$OFNS_DESC[df_arrest_ytd$OFNS_DESC == "KIDNAPPING"] <- "KIDNAPPING & RELATED OFFENSES"
df_arrest_ytd$OFNS_DESC[df_arrest_ytd$OFNS_DESC == "ADMINISTRATIVE CODES"] <- "ADMINISTRATIVE CODE"
```

```{r echo=TRUE, eval=FALSE}
arrests_loc_boro <- df_arrest_ytd %>% 
  filter(!OFNS_DESC %in% "") %>%
  group_by(OFNS_DESC, ARREST_BORO, Latitude, Longitude) %>%
  summarise(Total = n())

arrests_boro <- df_arrest_ytd %>% 
  filter(!OFNS_DESC %in% "") %>%
  group_by(OFNS_DESC, ARREST_BORO) %>%
  summarise(Total = n())

write.csv(arrests_boro, './datasets/filtered/arrests_by_offenses_and_borough.csv', row.names = FALSE)
```

Here to create a filtered dataset for interactive component we have identified summary information for offenses committed in each Borough. This data was also unstacked to achieve the current version of data.
 
## Human Trafficking

```{r echo=TRUE}
df_fbi_ht_by_states <- df_fbi_ht %>%
  group_by(STATE_NAME) %>%
  summarise(Total = n())

new_info <- data.frame( c("New York", "District of Columbia", "Pennsylvania", "California", "Alabama","Iowa"),  
                        c(0, 0, 0, 0, 0, 0))  
names(new_info) <- c("STATE_NAME", "Total")
df_fbi_ht_by_states <- rbind(df_fbi_ht_by_states, new_info)
```

## Shootings in NYC

### Pre-processing Data

```{r, echo=FALSE}

# df_shooting_ytd <- readr::read_csv("datasets/NYPD_Shooting_Incident_Data__Year_To_Date_.csv")
# df_shooting_old <- readr::read_csv("https://data.cityofnewyork.us/resource/833y-fsy8.csv")
```


```{r, echo=TRUE}
colnames(df_shooting_old)<- toupper(colnames(df_shooting_old))
colnames(df_shooting_ytd)<- toupper(colnames(df_shooting_ytd))
```

 - Column names in the old and new data have different case, converting to a common format for merging the 2 later.

```{r, echo=TRUE}
df_shooting_ytd$OCCUR_DATE <-as.Date(df_shooting_ytd$OCCUR_DATE, format = "%m/%d/%Y")
df_shooting_old$OCCUR_DATE <-as.Date(df_shooting_old$OCCUR_DATE, format = "%m/%d/%Y")

df_shooting_total<-bind_rows(df_shooting_ytd, df_shooting_old)
df_shooting_total$OCCUR_DATE <-as.Date(df_shooting_total$OCCUR_DATE, format = "%m/%d/%Y")
```

 - Converting the date columns in both to date class using format and then merging the 2 datasets in one.
 
### Feature Engineering

 - creating more features from the existing features to help 

```{r, echo=TRUE}
# extract year and month from date - #TODO: hour and dayofweek
df_shooting_total<-df_shooting_total %>%
  mutate(year = year(OCCUR_DATE), month = lubridate::month(OCCUR_DATE,label = TRUE), day = lubridate::wday(OCCUR_DATE,label = TRUE),
         murder = ifelse(STATISTICAL_MURDER_FLAG==TRUE, 1, 0))
```
 
  - Extracted year, month and day of the week
  - Calculated Percentage change in number of shootings every month.
  
```{r}

df_shooting_total$date<-as.Date(paste(df_shooting_total$year, df_shooting_total$month,"1",sep="-"), format = "%Y-%b-%d")
temp_df<- df_shooting_total %>% 
  # filter(year>=2015) %>%
  group_by(date, month) %>%
  summarize(total_shots= n())#%>%
  # mutate(monthly_change = (total_shots/shift(total_shots, n=1, fill = NA, type="lag") -1)*100)

temp_df$monthly_change<- ((temp_df$total_shots/shift(temp_df$total_shots, n=1, fill=NA, type="lag"))-1)

ggplot(temp_df, aes(x= date)) + 
  geom_bar(aes(y = total_shots, fill = factor(month)), stat="identity") +
  ggtitle("Shootings over time",
          subtitle = "Number of Shootings over time in NYC") +
  labs(y = "Number of Shootings", x="Date", fill="Month", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68")) +geom_vline(xintercept = covid_date, linetype="dashed", color = "blue") +
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - The data is very similar over the years and hence to keep our analysis latest, it is filtered to the last 5 years.
 
 
## Park Crime Data

The raw data we obtained was in the form of quarterly data from the year 2015 to 2020. In order to carry out effective analysis, we put together the individual datasets into one large .csv file. 

```{r eval=FALSE, echo=TRUE}
library(readxl)
library(httr)
q1_2015_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2015.xlsx', skip=4)
q2_2015_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2015.xlsx', skip=4)
q3_2015_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2015.xlsx', skip=4)
q4_2015_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2015.xlsx', skip=3)

q1_2016_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2016.xlsx', skip=3)
q2_2016_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2016.xlsx', skip=3)
q3_2016_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2016.xlsx', skip=3)
q4_2016_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2016.xlsx', skip=3)

q1_2017_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2017.xlsx', skip=3)
q2_2017_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2017.xlsx', skip=3)
q3_2017_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2017.xlsx', skip=3)
q4_2017_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2017.xlsx', skip=3)

q1_2018_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2018.xlsx', skip=3)
q2_2018_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2018.xlsx', skip=4)
q3_2018_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2018.xlsx', skip=3)
q4_2018_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2018.xlsx', skip=3)

q1_2019_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2019.xlsx', skip=3)
q2_2019_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2019.xlsx', skip=3)
q3_2019_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2019.xlsx', skip=3)
q4_2019_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2019.xlsx', skip=3)

q1_2020_park <-read_excel('datasets/parks/nyc-park-crime-stats-q1-2019.xlsx', skip=3)
q2_2020_park <-read_excel('datasets/parks/nyc-park-crime-stats-q2-2019.xlsx', skip=3)
q3_2020_park <-read_excel('datasets/parks/nyc-park-crime-stats-q3-2019.xlsx', skip=3)
q4_2020_park <-read_excel('datasets/parks/nyc-park-crime-stats-q4-2019.xlsx', skip=3)
names(q1_2020_park)
names(q2_2020_park)
names(q3_2020_park)
names(q4_2020_park)
```

Data has been read in and has to be converted to a large dataset, as a concatenation of the rows along with a quarter and year column.

```{r eval=FALSE, echo=TRUE}

q1_2015_park$year <- c(rep(2015, nrow(q1_2015_park)))
q1_2016_park$year <- c(rep(2016, nrow(q1_2016_park)))
q1_2017_park$year <- c(rep(2017, nrow(q1_2017_park)))
q1_2018_park$year <- c(rep(2018, nrow(q1_2018_park)))
q1_2019_park$year <- c(rep(2019, nrow(q1_2019_park)))
q1_2020_park$year <- c(rep(2020, nrow(q1_2020_park)))

q2_2015_park$year <- c(rep(2015, nrow(q2_2015_park)))
q2_2016_park$year <- c(rep(2016, nrow(q2_2016_park)))
q2_2017_park$year <- c(rep(2017, nrow(q2_2017_park)))
q2_2018_park$year <- c(rep(2018, nrow(q2_2018_park)))
q2_2019_park$year <- c(rep(2019, nrow(q2_2019_park)))
q2_2020_park$year <- c(rep(2020, nrow(q2_2020_park)))

q3_2015_park$year <- c(rep(2015, nrow(q3_2015_park)))
q3_2016_park$year <- c(rep(2016, nrow(q3_2016_park)))
q3_2017_park$year <- c(rep(2017, nrow(q3_2017_park)))
q3_2018_park$year <- c(rep(2018, nrow(q3_2018_park)))
q3_2019_park$year <- c(rep(2019, nrow(q3_2019_park)))
q3_2020_park$year <- c(rep(2020, nrow(q3_2020_park)))

q4_2015_park$year <- c(rep(2015, nrow(q4_2015_park)))
q4_2016_park$year <- c(rep(2016, nrow(q4_2016_park)))
q4_2017_park$year <- c(rep(2017, nrow(q4_2017_park)))
q4_2018_park$year <- c(rep(2018, nrow(q4_2018_park)))
q4_2019_park$year <- c(rep(2019, nrow(q4_2019_park)))
q4_2020_park$year <- c(rep(2020, nrow(q4_2020_park)))
```


```{r eval=FALSE, echo=TRUE}
q1_2015_park$quarter <- c(rep('q1', nrow(q1_2015_park)))
q1_2016_park$quarter <- c(rep('q1', nrow(q1_2016_park)))
q1_2017_park$quarter <- c(rep('q1', nrow(q1_2017_park)))
q1_2018_park$quarter <- c(rep('q1', nrow(q1_2018_park)))
q1_2019_park$quarter <- c(rep('q1', nrow(q1_2019_park)))
q1_2020_park$quarter <- c(rep('q1', nrow(q1_2020_park)))

q2_2015_park$quarter <- c(rep('q2', nrow(q2_2015_park)))
q2_2016_park$quarter <- c(rep('q2', nrow(q2_2016_park)))
q2_2017_park$quarter <- c(rep('q2', nrow(q2_2017_park)))
q2_2018_park$quarter <- c(rep('q2', nrow(q2_2018_park)))
q2_2019_park$quarter <- c(rep('q2', nrow(q2_2019_park)))
q2_2020_park$quarter <- c(rep('q2', nrow(q2_2020_park)))

q3_2015_park$quarter <- c(rep('q3', nrow(q3_2015_park)))
q3_2016_park$quarter <- c(rep('q3', nrow(q3_2016_park)))
q3_2017_park$quarter <- c(rep('q3', nrow(q3_2017_park)))
q3_2018_park$quarter <- c(rep('q3', nrow(q3_2018_park)))
q3_2019_park$quarter <- c(rep('q3', nrow(q3_2019_park)))
q3_2020_park$quarter <- c(rep('q3', nrow(q3_2020_park)))

q4_2015_park$quarter <- c(rep('q4', nrow(q4_2015_park)))
q4_2016_park$quarter <- c(rep('q4', nrow(q4_2016_park)))
q4_2017_park$quarter <- c(rep('q4', nrow(q4_2017_park)))
q4_2018_park$quarter <- c(rep('q4', nrow(q4_2018_park)))
q4_2019_park$quarter <- c(rep('q4', nrow(q4_2019_park)))
q4_2020_park$quarter <- c(rep('q4', nrow(q4_2020_park)))
```

```{r eval=FALSE, echo=TRUE}
colnames(q1_2015_park)<- toupper(colnames(q1_2015_park))
colnames(q2_2015_park)<- toupper(colnames(q2_2015_park))
colnames(q3_2015_park)<- toupper(colnames(q3_2015_park))
colnames(q4_2015_park)<- toupper(colnames(q4_2015_park))

colnames(q1_2016_park)<- toupper(colnames(q1_2016_park))
colnames(q2_2016_park)<- toupper(colnames(q2_2016_park))
colnames(q3_2016_park)<- toupper(colnames(q3_2016_park))
colnames(q4_2016_park)<- toupper(colnames(q4_2016_park))

colnames(q1_2017_park)<- toupper(colnames(q1_2017_park))
colnames(q2_2017_park)<- toupper(colnames(q2_2017_park))
colnames(q3_2017_park)<- toupper(colnames(q3_2017_park))
colnames(q4_2017_park)<- toupper(colnames(q4_2017_park))

colnames(q1_2018_park)<- toupper(colnames(q1_2018_park))
colnames(q2_2018_park)<- toupper(colnames(q2_2018_park))
colnames(q3_2018_park)<- toupper(colnames(q3_2018_park))
colnames(q4_2018_park)<- toupper(colnames(q4_2018_park))

colnames(q1_2019_park)<- toupper(colnames(q1_2019_park))
colnames(q2_2019_park)<- toupper(colnames(q2_2019_park))
colnames(q3_2019_park)<- toupper(colnames(q3_2019_park))
colnames(q4_2019_park)<- toupper(colnames(q4_2019_park))

colnames(q1_2020_park)<- toupper(colnames(q1_2020_park))
colnames(q2_2020_park)<- toupper(colnames(q2_2020_park))
colnames(q3_2020_park)<- toupper(colnames(q3_2020_park))
colnames(q4_2020_park)<- toupper(colnames(q4_2020_park))
```

Putting the dataframes together

```{r eval=FALSE, echo=TRUE}

park_data <- do.call("rbind", list(q1_2015_park, q2_2015_park ,q3_2015_park, q4_2015_park, q1_2016_park,q2_2016_park, q3_2016_park, q4_2016_park, q1_2017_park, q2_2017_park, q3_2017_park, q4_2017_park, q1_2018_park, q2_2018_park, q3_2018_park, q4_2018_park, q1_2019_park, q2_2019_park, q3_2019_park, q4_2019_park, q1_2020_park, q2_2020_park, q3_2020_park, q4_2020_park))

```

```{r eval=FALSE, echo=TRUE}
library(dplyr)
park_data <- park_data %>% filter(PARK != 'TOTAL')
```

Creating new csv file for output.

```{r eval=FALSE, echo=TRUE}
write.csv(park_data, './datasets/parks/finalpark.csv', row.names = FALSE)
```

Testing Dataset - Sanity Checks 

```{r eval=FALSE, echo=TRUE}
nrow(park_data)
```

## Hate Crime Data

County names are historical in this dataset, and hence have to be changed. 

```{r echo=TRUE}
df_hate <- read.csv("datasets/NYPD_Hate_Crimes.csv")
#View(df_hate)
df_hate$County[df_hate$County == "KINGS"] <- "BROOKLYN"
df_hate$County[df_hate$County == "NEW YORK"] <- "MANHATTAN"
df_hate$County[df_hate$County == "RICHMOND"] <- "STATEN ISLAND"
```
