==> 05-results.Rmd <==
# Results

## Shootings in NYC

### Monthly Trend in number of shootings

```{r}
covid_date<- as.Date("2020-04-01", format = "%Y-%m-%d")

df_shoot_grouped_NYC <- df_shooting_total %>%
  select(year, month, murder, VIC_AGE_GROUP, VIC_SEX, VIC_RACE) %>%
  group_by(year, month, murder, VIC_SEX, VIC_AGE_GROUP, VIC_RACE) %>%
  summarize(occurances = n())

df_shoot_grouped_NYC$date<-as.Date(paste(df_shoot_grouped_NYC$year, df_shoot_grouped_NYC$month,"1",sep="-"), format = "%Y-%b-%d")

df_shoot_grouped_BORO <- df_shooting_total %>%
  select(year, month, BORO, murder, VIC_AGE_GROUP, VIC_SEX, VIC_RACE) %>%
  group_by(year, month, murder, BORO, VIC_SEX, VIC_AGE_GROUP, VIC_RACE) %>%
  summarize(occurances = n())

df_shoot_grouped_BORO$BORO <- as.factor(df_shoot_grouped_BORO$BORO)

df_shoot_grouped_BORO$date<-as.Date(paste(df_shoot_grouped_BORO$year, df_shoot_grouped_BORO$month,"1",sep="-"), format = "%Y-%b-%d")

df_shoot_grouped_week <- df_shooting_total %>%
  select(OCCUR_DATE, day, STATISTICAL_MURDER_FLAG, VIC_AGE_GROUP, VIC_SEX, VIC_RACE) %>%
  group_by(OCCUR_DATE, day, STATISTICAL_MURDER_FLAG, VIC_SEX, VIC_AGE_GROUP, VIC_RACE) %>%
  summarize(occurances = n())

# df_shoot_grouped_week$date<-as.Date(paste(df_shoot_grouped_Week$year, df_shoot_grouped_Week$month,"1",sep="-"), format = "%Y-%b-%d")


```

```{r, fig.width=15,fig.height=11}
temp_df<- df_shoot_grouped_NYC %>% 
  filter(year>=2015) %>%
  group_by(date, month) %>%
  summarize(total_shots= sum(occurances))#%>%
  # mutate(monthly_change = (total_shots/shift(total_shots, n=1, fill = NA, type="lag") -1)*100)

temp_df$monthly_change<- ((temp_df$total_shots/shift(temp_df$total_shots, n=1, fill=NA, type="lag"))-1)

ggplot(temp_df, aes(x= date)) + 
  geom_bar(aes(y = total_shots, fill = factor(month)), stat="identity") +
  geom_text(aes( label = total_shots,
                  y= total_shots, vjust=-.5), stat= "identity") +
  ggtitle("Shootings over time",
          subtitle = "Number of Shootings over time in NYC") +
  labs(y = "Number of Shootings", x="Date", fill="Month", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68")) +geom_vline(xintercept = covid_date, linetype="dashed", color = "blue") +
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - The number of shootings have clearly gone up after covid (blue line).
 - Also, We see a trend increase in the number of shootings in the middle of the year which comes down in every winter.

```{r, fig.width=15,fig.height=11}

ggplot(temp_df, aes(x= date)) + 
  geom_bar(aes(y = monthly_change, fill = factor(month)), stat="identity") +
  geom_text(aes( label = scales::percent(monthly_change),
                  y= monthly_change ,vjust=-.5), stat= "identity", color="grey35", size=3) +
  scale_y_continuous(labels = scales::percent) +
  ggtitle("Change in shootings over time",
          subtitle = "Percentage change in number of Shootings over time") +
  labs(y = "Percentage Change", x="Date", fill="Month", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68")) +geom_vline(xintercept = covid_date, linetype="dashed", color = "blue") +
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - Almost every year the number of shootings rise rise sharply after march and then start decreasing in August.
 - Last year, April and may have the highest increase in any month over the pat years.  (possibly due to covid restrictions easing up, and people out of money, had to resort to other means) 

### Hot-spot neighbourhoods/AREAS 

#### Number of shootings in different borough every month

```{r}

temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015) %>%
  group_by(date,BORO) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

temp_df$BORO <- as.factor(temp_df$BORO)


ggplot(temp_df, aes(x = date, color = BORO, y = total_shots)) +
  geom_line() +
  ggtitle("Where to stay?!",
                       subtitle = "Number of Shootings in each borough by month") +
  labs(y = "Borough", x="Date", fill="Number of shootings", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme_pubr(legend="right")+
  geom_vline(xintercept = covid_date, linetype="dashed", color = "blue") +
  labs(color = "Borough", x="Date", y="Number of shootings", caption = "Source: NYC-open-Data::Shooting Incident Data")
```

 - This graph tells us, Staten island is at an all time low in the number of shootings, relatively.
 - Also, Brooklyn and Bronx have the most number of shootings which peaked during the month of july in 2020.

#### Percentage Change in number of shootings in different boroughs

```{r}
temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015) %>%
  group_by(BORO,date) %>%
  summarize(total_shots= sum(occurances)) %>%
  mutate(monthly_change = (total_shots/shift(total_shots, n=1, fill = NA, type="lag") -1)/10)

temp_df$month<- lubridate::month(temp_df$date,label = TRUE)
```

```{r, fig.width=10,fig.height=11}
ggplot(temp_df, aes(x= date)) + 
  geom_bar(aes(y = monthly_change, fill = factor(month)), stat="identity") +
  facet_grid(BORO~., scales = "free") +
  scale_y_continuous(labels = scales::percent)+
  ggtitle("Where to visit and When?!",
          subtitle = "Percentage change in number of Shootings in borough by month") +
  labs(y = "Percentage Change", x="Date", fill="Month", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +geom_vline(xintercept = covid_date, linetype="dashed", color = "blue") +
  theme(plot.caption = element_text(color = "grey68")) + 
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))

```
 
 - This faceted graph gives us a deeper dive in just the highest increase. Staten island and queens show an anomalous increase post covid. much higher than Brooklyn and Bronx. 
 - Also, in case of Staten island, the behavior is very erratic, hinting at a more loose control than other boroughs.


### Weekly Trend in number of shooting

```{r}
temp_df<- df_shoot_grouped_week %>% 
  group_by(day) %>%
  summarize(total_shots= mean(occurances))#%>%
  # mutate(monthly_change = (total_shots/shift(total_shots, n=1, fill = NA, type="lag") -1)*100)

temp_df$weekly_change<- ((temp_df$total_shots/shift(temp_df$total_shots, n=1, fill=NA, type="lag"))-1)
```


```{r}
ggplot(temp_df, aes(x= day, y = total_shots, fill = factor(day), label=sprintf("%0.2f", round(total_shots, digits = 2)))) + 
  geom_bar(stat="identity") +
  geom_text( vjust=-.5) +
  ggtitle("Does the day of the week matter?!",
          subtitle = "Average number of shootings on the day of the week") +
  labs(y = "Average shooting count", x="Day", fill="Weekday", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68"))+
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - Against the stereotypical belief people going out more on the weekends then weekdays and staying late. This graph shows that the day of the week is does not matter as much for a shooting incident.

### Trend in number of deaths
```{r, fig.width=10,fig.height=11}
temp_df<- df_shoot_grouped_NYC %>% 
  filter(year>=2015, murder==1) %>%
  group_by(date, month) %>%
  summarize(total_shots= sum(occurances))

ggplot(temp_df, aes(x= date)) + 
  geom_bar(aes(y = total_shots, fill = factor(month)), stat="identity") +
  geom_text(aes( label =total_shots,
                 y= total_shots, vjust=-.5), stat= "identity") +
  labs(y = "Total Deaths",x="Date", fill="Month Number",  caption = "Source: NYC-open-Data::Shooting Incident Data") +
  ggtitle("Number of deaths from shootings in NYC",
          subtitle = "Count of shootings that led to death over time") +geom_vline(xintercept = covid_date, linetype="dashed", color = "blue")+
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68"))+ 
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - The number of shootings lead to death increases from April to June and least every year in March. 
 - This follows from our previous finding of increase in the number of shootings during the same time of the year.

### Susceptible groups

#### Victims by Age

```{r}
temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015, VIC_AGE_GROUP!="UNKNOWN") %>%
  group_by(date,BORO, VIC_AGE_GROUP) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

temp_df$BORO <- as.factor(temp_df$BORO)

theme_heat <- theme_classic() +
  theme(axis.line = element_blank(),
        axis.ticks = element_blank())
```

```{r, fig.width=10,fig.height=11}
ggplot(temp_df, aes(x = date, y = VIC_AGE_GROUP)) +
  geom_tile(aes(fill = total_shots), color = "black") +
  theme_heat + ggtitle("Does Age matter?",
                       subtitle = "Number of Shootings in each borough for each age group over time") + scale_fill_gradient(low = "#cccccc", high = "#880808") +
  facet_grid(BORO~.) +
  labs(y = "Age Group", x="Date", fill="Number of shootings", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68")) +
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))
```

 - There are multiple deductions from the above graph:
   - The most widely affected age group is the 25-44 age group.
   - Brooklyn and Bronx have the highest shootings.

#### Victims by Gender

```{r}

temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015, VIC_SEX!="U") %>%
  group_by(date,BORO, VIC_SEX) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

ggplot(temp_df, aes(x = date, color = BORO, y = total_shots)) +
  geom_line() +
  ggtitle("Males and Females across NYC",
                       subtitle = "Number of Shootings in each borough by gender over time") +
  labs(y = "Number of shootings", x="Date", color="Borough", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme_pubr(legend="right") + facet_grid(VIC_SEX~., scales = "free", space="free_x") + 
  geom_vline(xintercept = covid_date, linetype="dashed", color = "blue")
```
  
 - The data on number of female victims in shootings is less but among them, brooklyn has the highest.
 - In general, we see a clear increase in the number of shootings post covid (blue-line)


```{r}
temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015, VIC_SEX!="U") %>%
  group_by(date, VIC_SEX) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

ggplot(temp_df, aes(x = date, color = VIC_SEX, y = total_shots)) +
  geom_line() +
  ggtitle("Males and Females victims in whole NYC",
                       subtitle = "Number of Shootings in NYC by gender over time") + scale_fill_gradient(low = "#cccccc", high = "#880808") +
  labs(y = "Number of shootings", x="Date", color="Borough", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  theme_pubr(legend="right")+
  geom_vline(xintercept = covid_date, linetype="dashed", color = "blue")
```
 
 - This visualization confirms that more men are involved in shootings than women.
 - Also, that the number has significantly risen post COVID.

#### Victims by Race

```{r, fig.width=10,fig.height=11}

temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015, VIC_RACE!="UNKNOWN", VIC_RACE!="AMERICAN INDIAN/ALASKAN NATIVE") %>%
  group_by(year,BORO, VIC_RACE) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

temp_df$year <- as.factor(temp_df$year)

ggplot(temp_df, aes(x = year, fill = BORO, y = total_shots)) +
  geom_bar(stat="identity", position = "dodge") +
  ggtitle("Are some areas bad for some race than others",
          subtitle = "Number of Shooting victims by race in each borough over time") +scale_fill_brewer(palette="Dark2") +
  labs(y = "Number of shootings", x="Year", fill="Borough", caption = "Source: NYC-open-Data::Shooting Incident Data") +
  facet_grid(VIC_RACE~., scales = "free", space="free_x") +
  theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68")) +
  theme(axis.text.y = element_text(size = rel(1.25)),axis.text.x = element_text(size = rel(1.25)), axis.title.x=element_text(size=rel(1.25)), axis.title.y=element_text(size=rel(1.25)))

```
 
 - We observe that different communities are effected more in some boroughs than others.
   - White people do not show up almost everywhere except in the occasional shootings in the Staten island borough.
   - Pacific Islanders are vicitimized in queens the highest.
   - White Hispanics and Black hispanics are targeted in Bronx.
   - Black in Brooklyn.

```{r}
temp_df<- df_shoot_grouped_BORO %>% 
  filter(year>=2015, VIC_RACE!="UNKNOWN", VIC_RACE!="AMERICAN INDIAN/ALASKAN NATIVE") %>%
  group_by(year, VIC_RACE) %>%
  summarize(total_shots= sum(occurances)) %>%
  ungroup()

temp_df$year <- as.factor(temp_df$year)

ggplot(temp_df, aes(x = year, fill = VIC_RACE, y = total_shots)) +
  geom_bar(stat="identity", position = "dodge") +
  ggtitle("Are any communities targetted?",
          subtitle = "Number of Shooting victims by race in NYC over time") +scale_fill_brewer(palette="Dark2") +
  labs(y = "Number of shootings", x="Year", fill="Victim Race", caption = "Source: NYC-open-Data::Shooting Incident Data") +
    theme(plot.title = element_text(face = "bold", size=rel(1.5))) +
  theme(plot.subtitle = element_text(face = "bold", color = "grey35")) +
  theme(plot.caption = element_text(color = "grey68"))
```

 - From just the numbers we see that the Black and White Hispanic people appear the most as victims across whole of New York.


```{r}
library(tidyverse)
library(ggplot2)

#df_fbi_ht <- readr::read_csv("datasets/HT_2013_2019.csv")

# fbiHT_df <- as.data.frame(df_fbi_ht)

# summary(df_fbi_ht)

theme_dotplot <- theme_bw(14) +
  theme(axis.text.y = element_text(size = rel(.75)),
        axis.ticks.y = element_blank(),
        axis.title.x = element_text(size = rel(.75)),
        panel.grid.major.x = element_blank(),
        panel.grid.major.y = element_line(size = 0.5),
        panel.grid.minor.x = element_blank())
```

## Human Trafficking

Note: The following regions were missing and their values are set to 0: Pennsylvania, California, Alabama, Iowa, New York, District of Columbia.

```{r}
df_fbi_ht_temp <- df_fbi_ht %>%
  group_by(STATE_NAME) %>%
  summarise(Total = n())

ggplot(df_fbi_ht_temp, aes(x = Total, y = reorder(STATE_NAME, Total))) +
  geom_point(color = "blue") +
  theme_dotplot +
  xlab("Total registered cases") +
  ylab("State") +
  ggtitle("Registered Human Trafficking cases by State (2013-19)")
```

```{r}
library(choroplethr)

df_fbi_ht_by_states <- df_fbi_ht %>%
  group_by(STATE_NAME) %>%
  summarise(Total = n())

new_info <- data.frame( c("New York", "District of Columbia", "Pennsylvania", "California", "Alabama","Iowa"),  
                        c(0, 0, 0, 0, 0, 0))  
names(new_info) <- c("STATE_NAME", "Total")
df_fbi_ht_by_states <- rbind(df_fbi_ht_by_states, new_info)

state_choropleth(df_fbi_ht_by_states %>% as.data.frame() %>% 
                   transmute(region = tolower(`STATE_NAME`), value = Total),
                 title = "Human Trafficking Cases By State (2013-19)",
                 legend = "Number of Cases")
```

From the above plots, we see that Texas and Minnesota record the highest number of human trafficking related arrests. A major reason for Texas to be one of the largest rings of labor and sex trafficking is due to its proximity to Mexico. This results in very high illegal immigrants being trafficked for sex, labor, drugs etc. -related purposes.

## NYPD Arrests

### Most Common Arrest Offenses

The first question we are looking to answer here is what are the most commons type of offenses as seen by NYPD. 

```{r fig.height=10, fig.width=10}
df_arrest_by_offense_type <- df_arrest_ytd %>% 
                              filter(!is.na(OFNS_DESC))%>%
                              group_by(OFNS_DESC) %>%
                              summarise(Total = n())

ggplot(df_arrest_by_offense_type, aes(x = Total, y = reorder(OFNS_DESC, Total))) +
  geom_point(color = "blue") +
  theme_dotplot +
  xlab("Total arrests") +
  ylab("Type of crime") +
  ggtitle("Most common crimes (as per NYPD Arrests Data)")
```


From the above plot, Assault, Felony Assault, Larceny, Drugs among others are the most common offenses seen in New York. These trends are consistent across all 5 boroughs.

#### Most Common Arrest Offenses By Sex

We also wish to identify the sex of the perpetrators committing these common offenses.

```{r fig.height=10, fig.width=10}
df_arrest_by_offense_type_perpsex <- df_arrest_ytd %>% 
  filter(!is.na(OFNS_DESC))%>%
  group_by(OFNS_DESC, PERP_SEX) %>%
  summarise(Total = n())

ggplot(df_arrest_by_offense_type_perpsex, aes(x = Total, y = reorder(OFNS_DESC, Total), color=PERP_SEX)) +
  geom_point() +
  theme_dotplot +
  xlab("Total arrests") +
  ylab("Type of crime") +
  ggtitle("Most common crimes by sex (as per NYPD Arrests Data)") +
  labs(color='Perp Sex')
```


The above plot follows a similar trend when it comes to the perpetrator's sex. However some offenses see an increase in the ratio of Female offenders to Men. These crimes include Criminal Mischief, Grand Larceny and offenses against public administration, among others.

### Assault Offenses by Borough

It is also important to identify crime rates by Borough to see which regions of New York are most affected by these crimes. Here we are focusing on assault related crimes since these are most prevalent in NYC.

```{r}
arrests_boro <- df_arrest_ytd %>% 
  filter(OFNS_DESC %in% "ASSAULT 3 & RELATED OFFENSES") %>%
  group_by(ARREST_BORO) %>%
  summarise(Total = n())

ggplot(arrests_boro, aes(x = fct_reorder(ARREST_BORO, Total), y = Total, fill=ARREST_BORO)) +
  geom_bar(stat = "identity") +
  ggtitle("Assault related offenses by Borough") +
  xlab("Boroughs") +
  ylab("Total arrests") +
  coord_flip() +
  labs(fill='Borough')
```


It is seen that Bronx and Brooklyn are most affected by Assault related offenses, whereas Staten Island is least affected. A basic cause for this could be assumed to be population of these boroughs but it is important to note here that while population decreases in the following order: Brooklyn, Queens, Manhattan, Bronx, Staten Island, this is not the trend of assault crimes.

### Offenses by Borough and Age Group

We try to further drill down among the NYC boroughs to understand if perpetrator age plays a role in these statistics.

```{r fig.width=12, fig.height=12}
arrests_boro_age <- df_arrest_ytd %>% 
  group_by(ARREST_BORO, AGE_GROUP) %>%
  summarise(Total = n())

ggplot(arrests_boro_age, aes(x = AGE_GROUP, y = Total, fill=AGE_GROUP)) +
  geom_bar(stat = "identity", position="dodge") +
  ggtitle("Assault related offenses by Borough") +
  xlab("Boroughs") +
  ylab("Total arrests") +
  facet_wrap(~ARREST_BORO) +
  labs(fill='Age Group')
```


The above plots show a similar trend across Boroughs with a majority of perpetrators belonging to the 25-44 age group. However it is important to note here that except for Manhattan, the number of perpetrators belonging to the 18-24 category are more than those belonging to 45-64 age group.

### Identifying race of perpetrators by Borough

It can also be valuable to know if there is a trend in perpetrator race for each borough.

```{r}
arrests_race <- df_arrest_ytd %>% 
  group_by(PERP_RACE, ARREST_BORO) %>%
  summarise(Total = n())

ggplot(arrests_race, aes(x = fct_reorder(ARREST_BORO, Total), y = Total, fill=PERP_RACE)) +
  geom_bar(stat = "identity") +
  ggtitle("Race of Perpetrators") +
  xlab("Race") +
  ylab("Total arrests") +
  labs(fill='Race')
```

The majority of perpetrator race seen here is Black, followed by White Hispanic. However it is important to note some other points as well:

 - State Island sees a similar distribution of Black as well as White perpetrators
 - Queens has a larger share of perpetrators belonging to Asian community
 - The Bronx has a very small percentage of perpetrators who are White. On the contrary, Black Hispanic is a more common group.

### Drugs based crimes in NYC

The following plots are used to understand density of drug-based crimes by location. 

```{r}
base_plot <- ggplot(df_arrest_ytd %>% 
                      filter(OFNS_DESC %in% "DANGEROUS DRUGS"), aes(x = Longitude, y = Latitude)) + 
  geom_point(aes(color = ARREST_BORO), size = 3, alpha = 0.5, position = "jitter") +
  xlab("Longitude") +
  ylab("Latitude") +
  ggtitle("Drugs Based Crimes, by Borough") +
  labs(color='Borough')

base_plot + theme_minimal()
```

```{r}
arrests_drugs_boro <- df_arrest_ytd %>% 
  filter(OFNS_DESC %in% "DANGEROUS DRUGS") %>% 
  group_by(Latitude, Longitude, ARREST_BORO) %>%
  summarise(Total = n())

ggplot(arrests_drugs_boro, aes(Longitude, Latitude)) + 
  geom_point(alpha=0.5, aes(color=ARREST_BORO)) + 
  geom_density_2d(bins = 5, size=1) +
  theme_minimal() +
  labs(color='Borough')
```

It is clear to us that Brooklyn and Bronx have more drug-related arrests, but most importantly Manhattan is where the trend is increasing and has high density.

### Sexual Crimes

An important aspect of crime analysis is understanding trends in sexual crimes and what areas are they more frequent in.

```{r}
sex_crimes <- ggplot(df_arrest_ytd %>% 
                      filter(OFNS_DESC %in% c("RAPE", "SEX CRIMES")), aes(x = Longitude, y = Latitude)) + 
  geom_point(aes(color = ARREST_BORO), size = 3, alpha = 0.5, position = "jitter") +
  xlab("Longitude") +
  ylab("Latitude") +
  ggtitle("Sexual Crimes and Rape") +
  labs(color='Borough')

sex_crimes + theme_minimal()
```

The above location based scatter plot shows the spread in location where sexual crimes tend to occur. To get a clearer picture, we plot the below density curve.

#### Sexual Crimes Frequency by Location

```{r}
sex_crimes2 <- ggplot(df_arrest_ytd %>% 
                       filter(OFNS_DESC %in% c("RAPE", "SEX CRIMES")), aes(x = Longitude, y = Latitude)) + 
  #geom_point(aes(color = ARREST_BORO), size = 3, alpha = 0.5, position = "jitter") +
  xlab("Longitude") +
  ylab("Latitude") +
  ggtitle("Sexual Crimes and Rape") + 
  stat_density2d(aes(fill = ..level..), alpha = .5,
                 geom = "polygon", data = df_arrest_ytd %>% 
                   filter(OFNS_DESC %in% c("RAPE", "SEX CRIMES")))+ 
  scale_fill_viridis_c() + 
  theme(legend.position = 'none')

sex_crimes2 + theme_minimal()
```

This plot identifies certain locations which tend to have a higher frequency of sexual crimes in NYC. Once again Brooklyn has the larger majority of sex crimes and rape related arrests, but now, we must note that Queens has a very high trend for the same. This can especially be noted for sexual crimes.

## Hate Crimes 

### Is there a certain county that is particularly bad when it comes to hate crimes? 

```{r}
library(ggplot2)
df_hate <- read.csv("datasets/NYPD_Hate_Crimes.csv")
hate3 <- ggplot(data=df_hate, aes(x=County, fill=factor(Complaint.Year.Number))) + geom_bar(data=df_hate) + labs(fill='Year') + ggtitle('Hate Crime Count per Borough - 2019/2020') + xlab('County/Borough') + ylab('Crime Conut')
hate3
```

Kings County (Brooklyn) had the highest total number of hate crimes over the years 2019 and 2020. This was followed by New York (Manhattan) and then Queens. 

The only county/borough to have experienced an upturn in hate crimes in 2020  was the Richmond county (Staten Island Borough). 

This is confirmed by articles such as *(Source: silive.com https://www.silive.com/crime-safety/2021/04/staten-island-was-the-only-borough-in-nyc-with-an-increase-in-hate-crimes-in-2020-why-where-were-they-crime-analysis.html)*

### What kinds of offense are particularly prevalent in different counties?

```{r}
hate1 <- ggplot(data=df_hate, aes(y=Offense.Category, fill=County)) + geom_bar(data=df_hate, alpha=0.9, position = position_stack(reverse = TRUE)) + xlab('Count') + ylab('Offense Category') + ggtitle('Offense type per Borough') + labs(fill='Borough')
hate1
```

The largest proportion of hate crimes stem from religious reasons, followed by race and sexual orientation. 

Kings County (Brooklyn) has a much larger proportion of these religious crimes. Interestingly, although the total number of hate crimes is greater in Kings County (Brooklyn) than New York County (Manhattan), the number of crimes stemming from Sexual Orientation and Race/Color is greater in New York County than Kings County. 


### Motives Count

```{r}
theme_dotplot <- theme_bw(14) +
    theme(axis.text.y = element_text(size = rel(.5)),
        axis.ticks.y = element_blank(),
        axis.title.x = element_text(size = rel(.75)),
        panel.grid.major.x = element_blank(),
        panel.grid.major.y = element_line(size = 1),
        panel.grid.minor.x = element_blank())

df_temp <- as.data.frame(table(df_hate$Bias.Motive.Description))
#View(df_temp)
hate2 <- ggplot(df_temp, aes(x = Freq, 
               y = reorder(Var1, Freq))) +
    geom_point(color = "blue") +
    theme_dotplot +
    xlab("Incident Count") +
    ylab("Motive Type") +
    ggtitle("Motive Type Counts") + xlim(0,400)
hate2
```


Anti- Jewish hate crimes seem to be the largest in number (by a large amount, almost 350 crimes) during this time period, followed by anti-homosexual (gay) and anti-black crimes. 


### What are the bias motives?

```{r fig.width=7}
hate4 <- ggplot(data=df_hate, aes(y=Bias.Motive.Description, fill=County)) + geom_bar(data=df_hate, alpha=0.9, position = position_stack(reverse = TRUE)) + xlab('Count') + ylab('Bias Motives') + ggtitle('Bias Motives split by Borough') + labs(fill='Borough')
hate4
```


As expected from the previous 2 graphs, Kings County (Brooklyn) has the highest number of anti-jewish crimes. Kings County has a disproportionately large number of these crimes, whereas other motive types like anti-black or anti-homosexual seem to be equal in the boroughs of Kings and New York (Manhattan).

These insights are confirmed by articles such as (https://www.cbsnews.com/news/nypd-to-increase-presence-in-brooklyn-after-string-of-anti-semitic-hate-crimes/ Source: CBS News)

### Has the distribution of crimes changed over the years?

```{r}
df_hate$Complaint.Year.Number <- as.factor(df_hate$Complaint.Year.Number)
hate5 <- ggplot(data=df_hate, aes(x=Complaint.Year.Number, fill=Offense.Category)) + geom_bar(data=df_hate) + ylab('Crime Count') + xlab('Complaint Year') + ggtitle('Distribution of Offense Categories per year') + labs(fill='Offense Category')
hate5
```

Across the city of New York, the number of hate crimes went down in 2020 (by almost 40%). Sexual orientation crimes decreased, whereas crimes based on race/color unfortunately went up in number. 

### Does the month have an influence on the number of crimes?

```{r}
df_temp <- as.data.frame(table(df_hate$Month.Number, df_hate$Complaint.Year.Number))

df_temp$new_col <- c(12:1)
#View(df_temp)
hate6 <- ggplot(df_temp, aes(x = Freq, 
               y = reorder(Var1, new_col), color=Var2)) +
    geom_point() +
    theme_dotplot +
    xlab("Incident Count") +
    ylab("Month Number") +
    ggtitle("Incident Counts by month") + labs(color='Year')
hate6
```

Across both years, it is visible that there is a dependence on the month when it comes to the number of crimes. 

However, for all months, 2020 had lower number of hate crimes than there were in 2019. The highest crime number for both years was in the month of March. 


### Year vs Bias Motive vs County

```{r fig.height=15, fig.width=15}
library(vcd)
hate7 <- vcd::doubledecker(Bias.Motive.Description ~ Complaint.Year.Number + County, data=df_hate, main='Analysing Motive with year and Borough') 
```

The insights from this mosaic plot corroborate our findings from the previous plots. 

### When do suspects get arrested? (If at all?)

```{r}
df_hate$date_diff <- as.Date(as.character(df_hate$Arrest.Date), format="%m/%d/%Y")-
                  as.Date(as.character(df_hate$Record.Create.Date), format="%m/%d/%Y")
hate8 <- ggplot(data=df_hate, mapping=aes(x=date_diff)) + geom_histogram( bins=20, color="#e9ecef", alpha=0.6, position = 'identity') + facet_wrap(~Complaint.Year.Number) + xlab('Difference - Days') + ylab('Frequency') + ggtitle('How many days till the suspect is arrested?')
hate8
```

The majority of arrests, if made, were made within 50 days since the report creation, for both the years 2019 and 2020. This indicates a heavy right-skew in the graphs. 

Only very few arrests were made beyond the 50-day mark in both years.

### How do the police view these?

```{r fig.height=15, fig.width=12}
hate9 <- ggplot(data=df_hate, aes(y=PD.Code.Description, fill=County)) + geom_bar(data=df_hate, alpha=0.9, position = position_stack(reverse = TRUE)) + xlab('PD - Code Description') + ylab('Crime Count') + ggtitle('How do the police classify these crimes and where are they most common? ')
hate9
```

Criminal Mischief, Aggravated Assault and Aggravated Harrassment were the most code of hate crimes committed.

## Drugs

```{r}
df_drugs <- readr::read_csv("datasets/arrests_national_drug.csv")
df_drugs <- as.data.frame(df_drugs)
#View(df_drugs)
```

### Arrests related to all drugs throughout US 
```{r}
library(ggplot2)
df_drugs$total_arrests <- df_drugs$total_arrests / 1000000
drug1 <- ggplot()+
  geom_line(data=df_drugs, aes(x = year, y = total_arrests, colour='blue')) +
  ggtitle("Arrests related to drugs across US across year") +
  labs(x = "year", y = "Number of Arrests (in millions)") + scale_color_discrete(name = "Arrest Count", labels = c("total arrests"))
drug1
```

#### Insights: all drugs

The number of drug related arrests across the United States peaked around the years 2005-2007 over the entire course of 1994 - 2016.

The lowest number of drug-related arrests were during 1994. The high number of drug related arrests coincides with the 

A large portion of the arrests related to upturn in drug-related arrests could be due to the raids detailed in articles such as:  https://www.justice.gov/archive/usao/nye/pr/2005/2005dec7.html. 
In addition, this could also be because of the setting up of the Digital Evidence Laboratory by the Drug Enforcement Administration. (Source: https://en.wikipedia.org/wiki/Drug_Enforcement_Administration)

### Arrests related to manufacture and possession
```{r}
df_drugs$total_manufacture <- df_drugs$total_manufacture/100000
df_drugs$total_possess <- df_drugs$total_possess/100000
drug2 <- ggplot()+
  geom_line(data=df_drugs, aes(x = year, y = total_manufacture, colour='blue')) +
  geom_line(data=df_drugs, aes(x = year, y = total_possess, colour='red'))+
  ggtitle("Arrests related to drug manufacture & possessions- US YoY") +
  labs(x = "year", y = "number of arrests (in 100k)") + scale_color_discrete(name = "Arrest Count", labels = c("Manufacture", "Possession"))
drug2
```

#### Insights: Manufacturing vs Possession

As seen in the graph, the number of arrests in relation to manufacturing of drugs has seen a downward trend, where the number of arrests in relation to possession of drugs has seen an upward trend. 

It is unclear whether the trends are due to a decrease in the manufacturing or due to other factors such as ineffectual police techniques. 

Similarly, the upward trend in the possession arrests of drugs could be due to either in increase in drug consumption or better police techniques to find such people to arrest, or both. 

### Arrests related to Manufacturing 

```{r}
df_drugs$opioid_manufacture <- df_drugs$opioid_manufacture/10000
df_drugs$marijuana_manufacture <- df_drugs$marijuana_manufacture/10000
df_drugs$synthetic_manufacture <- df_drugs$synthetic_manufacture/10000
df_drugs$other_manufacture <- df_drugs$other_manufacture/10000
drug3 <- ggplot()+
  geom_line(data=df_drugs, aes(x = year, y = opioid_manufacture, colour='blue')) +
  geom_line(data=df_drugs, aes(x = year, y = marijuana_manufacture, colour='red'))+
  geom_line(data=df_drugs, aes(x = year, y = synthetic_manufacture, colour='orange'))+
  geom_line(data=df_drugs, aes(x = year, y = other_manufacture, colour='green'))+
  ggtitle("Arrests related to drug manufacture -US YoY") +
  labs(x = "year", y = "Number of arrests (in 10^4)") + scale_color_discrete(name = "Arrest Count", labels = c("Opioid Manufacture", "Marijuana Manufacture", 'Synthetic Manufacture', 'Other Manufacture'))
drug3
```

#### Insights: Manufacturing Insights

The trends seen in the manufacture of these drugs is highly interesting. It is seen that the arrests related to opioid manufacture have seen a sharp downturn in the 1994-2016 period. 

In contrast, other drug-related arrests have been on the upturn, including marijuana, synthetic drugs and other drugs. 

### Arrests related to Possession

```{r}
df_drugs$opioid_possess <- df_drugs$opioid_possess/100000 
df_drugs$marijuana_possess <- df_drugs$marijuana_possess/100000
df_drugs$synthetic_possess <- df_drugs$synthetic_possess/100000
df_drugs$other_possess <- df_drugs$other_possess/100000
drug4 <- ggplot()+
  geom_line(data=df_drugs, aes(x = year, y = opioid_possess, colour='blue')) +
  geom_line(data=df_drugs, aes(x = year, y = marijuana_possess, colour='red'))+
  geom_line(data=df_drugs, aes(x = year, y = synthetic_possess, colour='orange'))+
  geom_line(data=df_drugs, aes(x = year, y = other_possess, colour='green'))+
  ggtitle("Arrests related to drug possesion-US YoY") +
  labs(x = "year", y = "number of arrests (in 100k)") + scale_color_discrete(name = "Arrest Count", labels = c("Opioid Possession", "Marijuana Possession", 'Synthetic Possession', 'Other Possession'))
drug4
```

#### Insights: Drug Possession Arrests

As can be seen from the scale of the graph, it is evident that the number of arrests due to drug possession are higher than the number of arrests in connection with drug manufacturing. 

In correlation with the manufacturing downtrend of opioid, the number of arrests for possession of opioid has also seen a downtrend. 

Again, similar to the manufacturing uptrend for other drugs such as marijuana or synthetic, the number of arrests have also been increasing during the 1993-2016 timeframe. 

### Arrests as a proportion of population

```{r}
#df_drugs <- subset(df_drugs, select=-c(prop))
df_drugs$population <- df_drugs$population/1000000
df_drugs$prop <- df_drugs$total_arrests/df_drugs$population
drug5 <- ggplot()+
  geom_line(data=df_drugs, aes(x = year, y = prop, colour='blue')) +
  ggtitle("Proportion of drug arrests compared to population -US YoY") +
  labs(x = "year", y = "number of arrests") + scale_color_discrete(name = "Arrest Count", labels = c("Population Proportion"))
drug5
```

#### Insights: Population Proportion

It is seen that there is a downward trend in the proportion of the us population arrested in connection to drugs. This could be due to the fact that the population growth has outstripped the number of drug manufacturers/possessors over this time frame (since it was seen that arrests have been on the increase during this timeframe).

## Park Crime Data

```{r}
df_parks <- read.csv('datasets/parks/finalpark.csv')
#View(df_parks)
```

### Crimes in Park by borough

```{r}
parks_subset <- subset(df_parks, select=c('BOROUGH', 'TOTAL', "YEAR"))
parks_subset <- filter(parks_subset, BOROUGH!='BROOKLYN/QUEENS')

parks_subset2 <- parks_subset %>% group_by(BOROUGH)
borough_total <- parks_subset2 %>% summarise(
  TOTAL = sum(TOTAL)
)
#View(parks_subset2)
ggplot() + geom_col(borough_total, mapping = aes(x=BOROUGH, y=TOTAL), fill='light blue', color='black') + labs(x='Borough', y='Total Crime Count') + ggtitle('Park Crime Count per Borough for 2015-2020')
```


Manhattan and Brooklyn parks have a particularly high count of total crimes over the 6 years. Staten Island parks have the lowest incidence of total crimes. 

### Parks with High Crime

*Any parks causing particularly large amounts of trouble (amongst parks with non-zero crime occurrence)?*

```{r fig.height=10}
parks_subset3 <- subset(df_parks, select=c('BOROUGH', 'TOTAL', 'PARK'))
#View(parks_subset3)
parks_subset3 <- filter(parks_subset3, TOTAL!=0)
parks_subset3 <- filter(parks_subset3, BOROUGH!='BROOKLYN/QUEENS')
ggplot(parks_subset3, aes(x = BOROUGH, y = TOTAL, label = PARK)) + 
  geom_boxplot() + xlab("") + theme_grey(16) + stat_summary(
    aes(label=round(stat(y), 1)),
    geom = "text", 
    fun.y = function(y) { o <- boxplot.stats(y)$out; if(length(o) == 0) NA else o },
    hjust = -1
  ) + labs(x='Borough', y='Total Crime Count per park') + ggtitle('Park Crime Count per Borough and Park ')

```

Thankfully, up to the 3rd quantile of parks ie 75% of parks in all boroughs have had 0 crimes over 6 years. Some parks in Manhattan and Brooklyn are particularly unsafe, however, and have particularly high incidence of crime in the years from 2015-2020. 

Three parks in Manhattan have caused problem in particular, with incidences of 71, 52, and 34 crimes each. While the highest number of crimes in a Brooklyn or Queens park is lower than the corresponding Manhattan number, the number of parks with greater than 20 crimes in total over the past 6 years in Brooklyn and Queens is greater than the number of Manhattan parks with greater than 20 crimes. 

### Park Size and Crime

*What effect does park size have on the occurrence of crime?*

```{r fig.width=8, fig.height=5}
park_grouped <- as.data.frame(df_parks) %>%
  group_by(YEAR, CATEGORY, BOROUGH) %>%
  summarise(Total = sum(TOTAL))
#View(park_grouped)
park_grouped <- filter(park_grouped, BOROUGH!='BROOKLYN/QUEENS')
ggplot(park_grouped, aes(x = Total, y = CATEGORY, fill=BOROUGH)) +
  geom_bar(position='dodge', stat = "identity") + ggtitle('Does park type have an effect on crime incidence?') + xlab('Crime Count') + ylab('Park Type')
```

We see here a very distinct dependence of crime on the park size. We see that crime occurs almost exclusively in the large parks.

Hence, clearly, parks that have larger sizes have higher incidence of crime. Large Manhattan and Brooklyn parks have the highest incidence of crime. Large Bronx and the Queens have identical incidence of crime. 

As parks get smaller, the incidence of crime drastically reduces across all boroughs. Playgrounds less than one acre have a few occurrences of crime, but interestingly, we see that the dependance on borough drastically reduces with crime occurences being low and almost equal across boroughs. 

This could suggest that criminals specifically target certain parks for organized or planned crime, something that does not occur in the smaller parks.

### Trends of different kinds of crime over the years

```{r fig.height=8, fig.width=10}
colnames(df_parks)[8] <- "FELONY_ASSAULT"
colnames(df_parks)[10] <- "GRAND_LARCENY"
colnames(df_parks)[11] <- "GRAND_LARCENY_MV"
park_grouped2 <- as.data.frame(df_parks) %>%
  group_by(YEAR,BOROUGH) %>%
  summarise(Total_murder = sum(MURDER), 
            Total_rape = sum(RAPE), 
            Total_robbery = sum(ROBBERY), 
            Total_fa = sum(FELONY_ASSAULT), 
            Total_burglary = sum(BURGLARY), 
            Total_grand = sum(GRAND_LARCENY), 
            Total_grandmv = sum(GRAND_LARCENY_MV), 
            total_total = sum(TOTAL))
park_grouped2 <- filter(park_grouped2, BOROUGH!='BROOKLYN/QUEENS')
ggplot() +
  geom_line(data=park_grouped2, mapping = aes(x=YEAR, y=Total_murder, color='murder')) + geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_rape, color='rape')) + geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_robbery, color='Robbery'))+geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_fa, color='Felony Assault'))+geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_burglary, color='Burglary'))+ geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_grand, color='Grand Larceny'))+ geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=Total_grandmv, color='Grand Larceny -Motor Vehicle'))+ facet_wrap(~BOROUGH) + ggtitle('Trends of crime types over the years') + xlab(' Borough') + ylab('Crime Count') + labs(color='Crime Type')
```


We see interesting patterns emerge in the type of park crimes occurring in different boroughs. 

In the Bronx, Robbery is the most common crime type, followed by Felony Assault and Grand Larceny. Brooklyn and Manhattan follow similar distributions to each other, with Felony Assault being the most frequent, followed by Robbery and Grand Larceny. In Queens, felony assault and robbery happen to be roughly equivalent in their occurrence. 

Interestingly, a lot of parks in the boroughs of Manhattan, Brooklyn and Queens saw an uptick in felony assault in 2016. Manhattan, however, had an extremely high occurrence of felony assault in 2017. 
Most boroughs have had either decreasing/stable park crime rates since 2017. 

Staten Island, even though it has the fewest parks, also has the smallest number of crimes by a large margin. 

### Time Analysis of Park Crimes

*Are the total number of park crimes in each borough increasing or decreasing over the years?*

```{r fig.height=5, fig.width=8}
ggplot() + geom_line(data= park_grouped2, mapping=aes(x=YEAR, y=total_total, linetype=BOROUGH)) + ggtitle('Total Crime Trend over the years') + labs(linetype='Borough')  + xlab('year') + ylab('Total Park Crimes')
```


Following upturns in the park crime rate in the year 2016, there has been a slight drop in the crime rates in Queens and the Bronx. 

However, Manhattan and Brooklyn saw an increase in park crime rates till 2017, followed by a slight increase in Manhattan and slight decrease in Brooklyn. Following this, the park crime occurrence has remained stable. 


### Further Analysis of Crime Type, Borough and Park Type

```{r fig.height=15, fig.width=15}
park_grouped4 <- as.data.frame(df_parks) %>%
  group_by(CATEGORY, BOROUGH) %>%
  summarise(Total_murder = sum(MURDER), 
            Total_rape = sum(RAPE), 
            Total_robbery = sum(ROBBERY), 
            Total_fa = sum(FELONY_ASSAULT), 
            Total_burglary = sum(BURGLARY), 
            Total_grand = sum(GRAND_LARCENY), 
            Total_grandmv = sum(GRAND_LARCENY_MV))
park_grouped4 <- filter(park_grouped4, BOROUGH!='BROOKLYN/QUEENS')
library(tidyr)
library(ggalluvial)
park_long <- gather(park_grouped4, type, count, Total_murder:Total_grandmv, factor_key=TRUE)
#View(park_long)
ggplot(park_long,
       aes(y = count, axis1 = BOROUGH, axis2 = CATEGORY)) +
  geom_alluvium(aes(fill = type), width = 1/12) +
  geom_stratum(width = 1/12, fill = "black", color = "grey") +
  geom_label(stat = "stratum", aes(label = after_stat(stratum))) +
  scale_x_discrete(limits = c("BOROUGH", "PARK CATEGORY"), expand = c(.05, .05)) + ggtitle('Alluvial Plot showing the type of crime in connection to Borough and Park Type') + ylab('Park Crime Count')
```

The insights from this alluvial diagram corroborate the findings from the previous plots. Manhattan and Brooklyn account for the most park crimes while large parks make up a very large percentage of the crime occurrence.
